requirements.txt
# requirements.txt 
google-genai
